refactor(knowledge_base): normalize wiki importance boundary to int
Align wiki importance semantics with filename and MCP contracts by\nnormalizing the WikiStore and inspection boundaries to positive\nintegers while keeping DB storage compatibility intact.\n\nUpdate wiki store and console-facing tests for the stricter\nimportance contract and record the completed slice in the\nrefactoring plan status.

+++ src/mai_gram/core/chat_inspection_service.py
@@ -30,7 +30,7 @@ class WikiEntryView:
 
     key: str
     value: str
-    importance: float
+    importance: int
 
 
 @dataclass(frozen=True, slots=True)
@@ -79,7 +79,7 @@ class ChatInspectionService:
                 WikiEntryView(
                     key=entry.key,
                     value=entry.value,
-                    importance=entry.importance,
+                    importance=int(entry.importance),
                 )
                 for entry in entries
             ],
